Fix wxSpinCtrlDouble properties min and max data type for XRC conversion

Use XrcFilter::Type::Float to preserve fractional digits.

+++ plugins/additional/additional.cpp
@@ -801,8 +801,8 @@ public:
         ObjectToXrcFilter xrc(GetLibrary(), obj, _("wxSpinCtrlDouble"), obj->GetPropertyAsString(_("name")));
         xrc.AddWindowProperties();
         xrc.AddProperty(_("initial"), _("value"), XrcFilter::Type::Text);
-        xrc.AddProperty(_("min"), _("min"), XrcFilter::Type::Integer);
-        xrc.AddProperty(_("max"), _("max"), XrcFilter::Type::Integer);
+        xrc.AddProperty(_("min"), _("min"), XrcFilter::Type::Float);
+        xrc.AddProperty(_("max"), _("max"), XrcFilter::Type::Float);
         xrc.AddProperty(_("inc"), _("inc"), XrcFilter::Type::Float);
         xrc.AddProperty(_("digits"), _("digits"), XrcFilter::Type::Integer);
         return xrc.GetXrcObject();
@@ -814,8 +814,8 @@ public:
         filter.AddWindowProperties();
         filter.AddProperty(_("value"), _("value"), XrcFilter::Type::Text);
         filter.AddProperty(_("value"), _("initial"), XrcFilter::Type::Text);
-        filter.AddProperty(_("min"), _("min"), XrcFilter::Type::Integer);
-        filter.AddProperty(_("max"), _("max"), XrcFilter::Type::Integer);
+        filter.AddProperty(_("min"), _("min"), XrcFilter::Type::Float);
+        filter.AddProperty(_("max"), _("max"), XrcFilter::Type::Float);
         filter.AddProperty(_("inc"), _("inc"), XrcFilter::Type::Float);
         filter.AddProperty(_("digits"), _("digits"), XrcFilter::Type::Integer);
